+++ tests/test_plotgrid.py
@@ -1,15 +1,18 @@
 from pytest import raises
 from spb import (
     MB, PB, BB, KB, plotgrid, plot, plot3d, plot_contour, plot_vector,
-    plot_polar
+    plot_polar, PlotGrid, plot_complex, plot_parametric,
+    plot3d_parametric_line, plot_contour
 )
 from spb.plotgrid import _nrows_ncols
-from sympy import symbols, sin, cos, tan, exp, pi
+from sympy import symbols, sin, cos, tan, exp, pi, Piecewise
 import matplotlib
 import matplotlib.pyplot as plt
 import mpl_toolkits
 import panel as pn
-
+import numpy as np
+import os
+from tempfile import TemporaryDirectory
 
 class KBchild1(KB):
     def _get_mode(self):
@@ -26,56 +29,58 @@ def test_nrows_ncols():
     assert nr == 5 and nc == 1
     nr, nc = _nrows_ncols(0, -1, 5)
     assert nr == 5 and nc == 1
-    nr, nc = _nrows_ncols(1, 1, 5)
-    assert nr == 5 and nc == 1
 
     # one row
     nr, nc = _nrows_ncols(1, -1, 5)
     assert nr == 1 and nc == 5
     nr, nc = _nrows_ncols(1, 0, 5)
     assert nr == 1 and nc == 5
+    nr, nc = _nrows_ncols(1, 1, 5)
+    assert nr == 1 and nc == 5
+    nr, nc = _nrows_ncols(1, 2, 5)
+    assert nr == 1 and nc == 5
 
     # not enough grid-elements to plot all plots: keep adding rows
-    nr, nc = _nrows_ncols(1, 2, 5)
-    assert nr == 3 and nc == 2
     nr, nc = _nrows_ncols(2, 2, 5)
     assert nr == 3 and nc == 2
 
     # enough grid-elements: do not modify nr, nc
     nr, nc = _nrows_ncols(2, 2, 4)
     assert nr == 2 and nc == 2
-
     nr, nc = _nrows_ncols(3, 2, 5)
     assert nr == 3 and nc == 2
 
 
 def test_empty_plotgrid():
     p = plotgrid(show=False)
-    assert isinstance(p, plt.Figure)
+    assert isinstance(p, PlotGrid)
+    assert isinstance(p.fig, plt.Figure)
 
 
 def test_plotgrid_mode_1():
     x, y, z = symbols("x, y, z")
+    options = dict(adaptive=False, n=10, backend=MB, show=False)
 
     # all plots with MatplotlibBackend: combine them into a matplotlib figure
-    p1 = plot(cos(x), (x, -5, 5), adaptive=False, n=100,
-        backend=MB, show=False, ylabel="a", use_latex=True)
-    p2 = plot(sin(x), (x, -7, 7), adaptive=False, n=100,
-        backend=MB, show=False, ylabel="b", use_latex=False)
-    p3 = plot(tan(x), (x, -10, 10), adaptive=False, n=100,
-        backend=MB, show=False, ylabel="c", use_latex=False)
+    p1 = plot(cos(x), (x, -5, 5), ylabel="a", use_latex=True, **options)
+    p2 = plot(sin(x), (x, -7, 7), ylabel="b", use_latex=False, **options)
+    p3 = plot(tan(x), (x, -10, 10), ylabel="c", use_latex=False, **options)
     p = plotgrid(p1, p2, p3, show=False)
-    assert isinstance(p, plt.Figure)
-    assert len(p.axes) == 3
-    assert p.axes[0].get_xlabel() == "$x$" and p.axes[0].get_ylabel() == "a"
-    assert len(p.axes[0].get_lines()) == 1
-    assert p.axes[0].get_lines()[0].get_label() == "$\\cos{\\left(x \\right)}$"
-    assert p.axes[1].get_xlabel() == "x" and p.axes[1].get_ylabel() == "b"
-    assert len(p.axes[1].get_lines()) == 1
-    assert p.axes[1].get_lines()[0].get_label() == "sin(x)"
-    assert p.axes[2].get_xlabel() == "x" and p.axes[2].get_ylabel() == "c"
-    assert len(p.axes[2].get_lines()) == 1
-    assert p.axes[2].get_lines()[0].get_label() == "tan(x)"
+    assert isinstance(p, PlotGrid)
+    assert isinstance(p.fig, plt.Figure)
+    assert len(p.fig.axes) == 3
+    assert p.fig.axes[0].get_xlabel() == "$x$"
+    assert p.fig.axes[0].get_ylabel() == "a"
+    assert len(p.fig.axes[0].get_lines()) == 1
+    assert p.fig.axes[0].get_lines()[0].get_label() == "$\\cos{\\left(x \\right)}$"
+    assert p.fig.axes[1].get_xlabel() == "x"
+    assert p.fig.axes[1].get_ylabel() == "b"
+    assert len(p.fig.axes[1].get_lines()) == 1
+    assert p.fig.axes[1].get_lines()[0].get_label() == "sin(x)"
+    assert p.fig.axes[2].get_xlabel() == "x"
+    assert p.fig.axes[2].get_ylabel() == "c"
+    assert len(p.fig.axes[2].get_lines()) == 1
+    assert p.fig.axes[2].get_lines()[0].get_label() == "tan(x)"
 
     # no errors are raised when the number of plots is less than the number
     # of grid-cells
@@ -84,34 +89,36 @@ def test_plotgrid_mode_1():
     # everything works fine when including 3d plots
     p1 = plot3d(cos(x**2 + y**2), (x, -2, 2), (y, -2, 2),
         backend=MB, n1=20, n2=20, show=False, use_latex=False)
-    p2 = plot(sin(x), cos(x), (x, -7, 7), adaptive=False, n=100,
-        backend=MB, show=False, use_latex=False)
+    p2 = plot(sin(x), cos(x), (x, -7, 7), use_latex=False, **options)
     p = plotgrid(p1, p2, nc=2, show=False)
-    assert isinstance(p, plt.Figure)
-    assert len(p.axes) == 2
-    assert isinstance(p.axes[0], mpl_toolkits.mplot3d.Axes3D)
-    assert not isinstance(p.axes[1], mpl_toolkits.mplot3d.Axes3D)
-    assert p.axes[0].get_xlabel() == "x" and p.axes[0].get_ylabel() == "y" and p.axes[0].get_zlabel() == "f(x, y)"
-    assert len(p.axes[0].collections) == 1
-    assert p.axes[1].get_xlabel() == "x" and p.axes[1].get_ylabel() == "f(x)"
-    assert len(p.axes[1].get_lines()) == 2
+    assert isinstance(p, PlotGrid)
+    assert isinstance(p.fig, plt.Figure)
+    assert len(p.fig.axes) == 2
+    assert isinstance(p.fig.axes[0], mpl_toolkits.mplot3d.Axes3D)
+    assert not isinstance(p.fig.axes[1], mpl_toolkits.mplot3d.Axes3D)
+    assert p.fig.axes[0].get_xlabel() == "x"
+    assert p.fig.axes[0].get_ylabel() == "y"
+    assert p.fig.axes[0].get_zlabel() == "f(x, y)"
+    assert len(p.fig.axes[0].collections) == 1
+    assert p.fig.axes[1].get_xlabel() == "x"
+    assert p.fig.axes[1].get_ylabel() == "f(x)"
+    assert len(p.fig.axes[1].get_lines()) == 2
 
     # mix different backends
-    p1 = plot(cos(x), (x, -3, 3), adaptive=False, n=100,
-        backend=MB, show=False)
+    p1 = plot(cos(x), (x, -3, 3), **options)
     p2 = plot_contour(cos(x**2 + y**2), (x, -3, 3), (y, -3, 3),
         backend=PB, n1=20, n2=20, show=False)
     p3 = plot3d(cos(x**2 + y**2), (x, -3, 3), (y, -3, 3),
         backend=KBchild1, n1=20, n2=20, show=False)
     p4 = plot_vector([-y, x], (x, -5, 5), (y, -5, 5), backend=BB, show=False)
 
-    p = plotgrid(p1, p2, p3, p4, nr=2, nc=2)
-    assert isinstance(p, pn.GridSpec)
-    assert p.ncols == 2 and p.nrows == 2
-    assert isinstance(p.objects[(0, 0, 1, 1)], pn.pane.plot.Matplotlib)
-    assert isinstance(p.objects[(0, 1, 1, 2)], pn.pane.plotly.Plotly)
-    assert isinstance(p.objects[(1, 0, 2, 1)], pn.pane.ipywidget.IPyWidget)
-    assert isinstance(p.objects[(1, 1, 2, 2)], pn.pane.plot.Bokeh)
+    p = plotgrid(p1, p2, p3, p4, nr=2, nc=2, show=False)
+    assert isinstance(p.fig, pn.GridSpec)
+    assert p.ncolumns == 2 and p.nrows == 2
+    assert isinstance(p.fig.objects[(0, 0, 1, 1)], pn.pane.plot.Matplotlib)
+    assert isinstance(p.fig.objects[(0, 1, 1, 2)], pn.pane.plotly.Plotly)
+    assert isinstance(p.fig.objects[(1, 0, 2, 1)], pn.pane.ipywidget.IPyWidget)
+    assert isinstance(p.fig.objects[(1, 1, 2, 2)], pn.pane.plot.Bokeh)
 
 
 def test_plotgrid_mode_2():
@@ -128,16 +135,14 @@ def test_plotgrid_mode_2():
         assert len(ax.collections) == 1
 
     # all plots are instances of MatplotlibBackend
-    p1 = plot(exp(x), adaptive=False, n=100, backend=MB,
+    options = dict(adaptive=False, n=100, backend=MB,
         show=False, use_latex=False)
-    p2 = plot(sin(x), adaptive=False, n=100, backend=MB,
-        show=False, use_latex=False)
-    p3 = plot(tan(x), backend=MB, show=False, adaptive=False,
-        detect_poles=True, eps=0.1, ylim=(-5, 5), use_latex=False)
-    p4 = plot(cos(x), adaptive=False, n=100, backend=MB,
-        show=False, use_latex=False)
-    p5 = plot3d(cos(x**2 + y**2), (x, -2, 2), (y, -2, 2),
-        backend=MB, n1=20, n2=20, show=False, use_latex=False)
+    p1 = plot(exp(x), **options)
+    p2 = plot(sin(x), **options)
+    p3 = plot(tan(x), detect_poles=True, eps=0.1, ylim=(-5, 5), **options)
+    p4 = plot(cos(x), **options)
+    options["n"] = 20
+    p5 = plot3d(cos(x**2 + y**2), (x, -2, 2), (y, -2, 2), **options)
 
     # gs is not a dictionary
     raises(TypeError, lambda: plotgrid(p1, p2, p3, p4, p5, gs=1, show=False))
@@ -155,26 +160,26 @@ def test_plotgrid_mode_2():
     }
     p = plotgrid(gs=mapping, show=False)
 
-    assert isinstance(p, plt.Figure)
-    assert len(p.axes) == 5
-    assert isinstance(p.axes[-1], mpl_toolkits.mplot3d.Axes3D)
-    test_line_axes(p.axes[0])
-    test_line_axes(p.axes[1])
-    test_line_axes(p.axes[2])
-    test_line_axes(p.axes[3])
-    test_3d_axes(p.axes[4])
+    assert isinstance(p, PlotGrid)
+    assert isinstance(p.fig, plt.Figure)
+    assert len(p.fig.axes) == 5
+    assert isinstance(p.fig.axes[-1], mpl_toolkits.mplot3d.Axes3D)
+    test_line_axes(p.fig.axes[0])
+    test_line_axes(p.fig.axes[1])
+    test_line_axes(p.fig.axes[2])
+    test_line_axes(p.fig.axes[3])
+    test_3d_axes(p.fig.axes[4])
 
     # Mixture of different backends
-    p1 = plot(exp(x), adaptive=False, n=100, backend=MB,
+    options = dict(adaptive=False, n=100,
         show=False, use_latex=False)
-    p2 = plot(sin(x), adaptive=False, n=100, backend=PB,
-        show=False, use_latex=False)
-    p3 = plot(tan(x), backend=MB, show=False, adaptive=False,
-        detect_poles=True, eps=0.1, ylim=(-5, 5), use_latex=False)
-    p4 = plot(cos(x), adaptive=False, n=100, backend=BB,
-        show=False, use_latex=False)
+    p1 = plot(exp(x), backend=MB, **options)
+    p2 = plot(sin(x), backend=PB, **options)
+    p3 = plot(tan(x), backend=MB, detect_poles=True, eps=0.1, ylim=(-5, 5),
+        **options)
+    p4 = plot(cos(x), backend=BB, **options)
     p5 = plot3d(cos(x**2 + y**2), (x, -2, 2), (y, -2, 2),
-        backend=KBchild1, n1=20, n2=20, show=False, use_latex=False)
+        backend=KBchild1, n=20, show=False, use_latex=False)
 
     gs = GridSpec(3, 3)
     mapping = {
@@ -184,26 +189,28 @@ def test_plotgrid_mode_2():
         gs[2:, 1:]: p4,
         gs[0:2, 1:]: p5,
     }
-    p = plotgrid(gs=mapping)
-    assert isinstance(p, pn.GridSpec)
-    assert p.nrows == p.ncols == 3
-    assert len(p.objects) == 5
-    assert isinstance(p.objects[(0, 0, 1, 1)], pn.pane.plot.Matplotlib)
-    assert isinstance(p.objects[(1, 0, 2, 1)], pn.pane.plotly.Plotly)
-    assert isinstance(p.objects[(2, 0, 3, 1)], pn.pane.plot.Matplotlib)
-    assert isinstance(p.objects[(2, 1, 3, 3)], pn.pane.plot.Bokeh)
-    assert isinstance(p.objects[(0, 1, 2, 3)], pn.pane.ipywidget.IPyWidget)
+    p = plotgrid(gs=mapping, show=False)
+    assert isinstance(p, PlotGrid)
+    assert isinstance(p.fig, pn.GridSpec)
+    assert p.fig.nrows == p.fig.ncols == 3
+    assert len(p.fig.objects) == 5
+    assert isinstance(p.fig.objects[(0, 0, 1, 1)], pn.pane.plot.Matplotlib)
+    assert isinstance(p.fig.objects[(1, 0, 2, 1)], pn.pane.plotly.Plotly)
+    assert isinstance(p.fig.objects[(2, 0, 3, 1)], pn.pane.plot.Matplotlib)
+    assert isinstance(p.fig.objects[(2, 1, 3, 3)], pn.pane.plot.Bokeh)
+    assert isinstance(p.fig.objects[(0, 1, 2, 3)], pn.pane.ipywidget.IPyWidget)
 
 
 def test_panel_kw():
     x = symbols("x")
-    p1 = plot(sin(x), adaptive=False, n=100, backend=MB, show=False)
-    p2 = plot(tan(x), adaptive=False, n=100, backend=PB, show=False)
-    p3 = plot(exp(-x), adaptive=False, n=100, backend=BB, show=False)
+    options = dict(adaptive=False, n=100, show=False)
+    p1 = plot(sin(x), backend=MB, **options)
+    p2 = plot(tan(x), backend=PB, **options)
+    p3 = plot(exp(-x), backend=BB, **options)
     pg1 = plotgrid(p1, p2, p3, nr=1, nc=3, show=False)
     pg2 = plotgrid(p1, p2, p3, nr=1, nc=3, show=False,
         panel_kw=dict(sizing_mode="stretch_width", height=250))
-    assert (pg1.height != pg2.height) and (pg2.height == 250)
+    assert (pg1.fig.height != pg2.fig.height) and (pg2.fig.height == 250)
 
 
 def test_plotgrid_mode_1_polar_axis():
@@ -217,5 +224,103 @@ def test_plotgrid_mode_1_polar_axis():
         r, (theta, 0, 2 * pi), show=False, backend=MB,
         polar_axis=pa, ylim=yl)
     pg = plotgrid(p(True, (0, 3)), p(), nr=1, nc=-1, show=False)
-    assert isinstance(pg.axes[0], matplotlib.projections.polar.PolarAxes)
-    assert not isinstance(pg.axes[1], matplotlib.projections.polar.PolarAxes)
+    assert isinstance(pg.fig.axes[0], matplotlib.projections.polar.PolarAxes)
+    assert not isinstance(pg.fig.axes[1], matplotlib.projections.polar.PolarAxes)
+
+
+def test_plot_size():
+    # verify that the `size` keyword argument works fine
+
+    x = symbols("x")
+    options = dict(adaptive=False, n=100, backend=MB, show=False)
+    p1 = plot(sin(x), **options)
+    p2 = plot(cos(x), **options)
+    p = plotgrid(p1, p2, size=(5, 7.5), show=False)
+    assert np.allclose(p.fig.get_size_inches(), [5, 7.5])
+
+    options = dict(adaptive=False, n=100, backend=PB, show=False)
+    p1 = plot(sin(x), **options)
+    p2 = plot(cos(x), **options)
+    p = plotgrid(p1, p2, size=(400, 600), show=False)
+    assert p.fig.width == 400
+    assert p.fig.height == 600
+
+
+def test_imagegrid():
+    # verify that matplotlib's ImageGrid axis works fine
+
+    x = symbols("x")
+    options = dict(backend=MB, show=False, n=10)
+    p1 = plot_complex(sin(x), (x, -2-2j, 2+2j), **options)
+    p2 = plot_complex(cos(x), (x, -2-2j, 2+2j), **options)
+    p = plotgrid(p1, p2, nr=1, imagegrid=True, show=False)
+    assert len(p.fig.axes) == 4
+    assert p.fig.axes[0].get_xlabel() == "Re"
+    assert p.fig.axes[1].get_xlabel() == "Re"
+    assert p.fig.axes[2].get_ylabel() == "Argument"
+
+    p = plotgrid(p1, p2, nr=1, imagegrid=False, show=False)
+    assert len(p.fig.axes) == 4
+    assert p.fig.axes[0].get_xlabel() == "Re"
+    assert p.fig.axes[1].get_ylabel() == "Argument"
+    assert p.fig.axes[2].get_xlabel() == "Re"
+    assert p.fig.axes[3].get_ylabel() == "Argument"
+
+
+def test_save():
+    x, y = symbols("x, y")
+
+    def _create_plots(B):
+        options = dict(adaptive=False, n=10, show=False, backend=B)
+        p1 = plot(x, **options)
+        p2 = plot_parametric((sin(x), cos(x)), (x, sin(x)), **options)
+        p3 = plot_parametric(
+            cos(x), sin(x), **options)
+        p4 = plot3d_parametric_line(sin(x), cos(x), x, **options)
+        p5 = plot(cos(x), (x, -pi, pi), **options)
+        p5[0].color_func = lambda a: a
+        p6 = plot(Piecewise((1, x > 0), (0, True)), (x, -1, 1), **options)
+        p7 = plot_contour(
+            (x**2 + y**2, (x, -5, 5), (y, -5, 5)),
+            (x**3 + y**3, (x, -3, 3), (y, -3, 3)), **options)
+        return p1, p2, p3, p4, p5, p6, p7
+
+    with TemporaryDirectory(prefix='sympy_') as tmpdir:
+        # matplotlib figures
+        p1, p2, p3, p4, p5, p6, p7 = _create_plots(MB)
+        
+        # symmetric grid
+        p = PlotGrid(2, 2, p1, p2, p3, p4, show=False)
+        filename = 'test_grid1.png'
+        p.save(os.path.join(tmpdir, filename))
+        p.close()
+
+        # grid size greater than the number of subplots
+        p = PlotGrid(3, 4, p1, p2, p3, p4, show=False)
+        filename = 'test_grid2.png'
+        p.save(os.path.join(tmpdir, filename))
+        p.close()
+
+        # unsymmetric grid (subplots in one line)
+        p = PlotGrid(1, 3, p5, p6, p7, show=False)
+        filename = 'test_grid3.png'
+        p.save(os.path.join(tmpdir, filename))
+        p.close()
+
+
+        # holoviz's panel objects
+        p1, p2, p3, p4, p5, p6, p7 = _create_plots(PB)
+
+        # symmetric grid
+        p = PlotGrid(2, 2, p1, p2, p3, p4, show=False)
+        p.save("test_1.html")
+
+        # grid size greater than the number of subplots
+        p = PlotGrid(3, 4, p1, p2, p3, p4, show=False)
+        p.save("test_2.html")
+
+        # unsymmetric grid (subplots in one line)
+        p = PlotGrid(1, 3, p5, p6, p7, show=False)
+        p.save("test_3.html")
+        
+